[dev sign] rename HAL_Snprintf declaration

+++ src/dev_sign/dev_sign_wrapper.h
@@ -5,4 +5,4 @@
 
 void *HAL_Malloc(uint32_t size);
 void HAL_Free(void *ptr);
-int8_t HAL_Snprintf(char *str, const int len, const char *fmt, ...);
+int HAL_Snprintf(char *str, const int len, const char *fmt, ...);
